refactor: inline in the clone utility
